+++ Java/DormitoryNotebooksJava/src/main/java/kaczuch/master_thesis/repositories/UserRepository.java
@@ -6,11 +6,11 @@ import kaczuch.master_thesis.model.User;
 import java.util.List;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 	User findByEmail (String email);
 
-	List<User> findByOrganizations_Id(Long organizationId);
+	List<User> findByOrganizations_Id(Integer organizationId);
 
-	List<User> findByDorms_Id(Long dormId);
+	List<User> findByDorms_Id(Integer dormId);
 
 }
